move fake API from test_ibmq_job_states
