+++ qiskit/test/mock.py
@@ -7,11 +7,13 @@
 
 
 """
-Supporting fake, stubs and mocking classes.
+Utilities for mocking the IBMQ provider, including job responses and backends.
 
-The module includes, among other, a dummy backend simulator. The purpose of
-this class is to create a Simulator that we can trick for testing purposes:
+The module includes, among other, a dummy backend. The purpose of
+this class is to create backends that we can trick for testing purposes:
 testing local timeouts, arbitrary responses or behavior, etc.
+
+The mock devices are mainly for testing the compiler.
 """
 
 import uuid
@@ -32,23 +34,35 @@ from qiskit.providers.baseprovider import BaseProvider
 logger = logging.getLogger(__name__)
 
 
-class DummyProvider(BaseProvider):
-    """Dummy provider just for testing purposes."""
+class FakeProvider(BaseProvider):
+    """Dummy provider just for testing purposes.
+
+    Only filtering backends by name is implemented.
+    """
 
     def get_backend(self, name=None, **kwargs):
-        return self._backend
+        backend = self._backends[0]
+        if name:
+            filtered_backends = [backend for backend in self._backends
+                                 if backend.name() == name]
+            if not filtered_backends:
+                raise QiskitBackendNotFoundError()
+            else:
+                backend = filtered_backends[0]
+        return backend
 
     def backends(self, name=None, **kwargs):
-        return [self._backend]
+        return self._backends
 
     def __init__(self):
-        self._backend = DummySimulator()
-
+        self._backends = [FakeQasmSimulator(),
+                          FakeTenerife(),
+                          FakeRueschlikon()]
         super().__init__()
 
 
-class DummySimulator(BaseBackend):
-    """ This is Dummy backend simulator just for testing purposes """
+class FakeBackend(BaseBackend):
+    """This is a dummy backend just for testing purposes."""
 
     DEFAULT_CONFIGURATION = {
         'name': 'local_dummy_simulator',
@@ -88,21 +102,123 @@ class DummySimulator(BaseBackend):
 
     def run(self, qobj):
         job_id = str(uuid.uuid4())
-        job = DummyJob(self.run_job, qobj, job_id, self)
+        job = FakeJob(self.run_job, qobj, job_id, self)
         job.submit()
         return job
 
     # pylint: disable=unused-argument
     def run_job(self, job_id, qobj):
-        """ Main dummy simulator loop """
+        """Main dummy run loop"""
         time.sleep(self.time_alive)
 
         return Result.from_dict(
             {'job_id': job_id, 'result': [], 'status': 'COMPLETED'})
 
 
-class DummyJob(BaseJob):
-    """Dummy simulator job"""
+class FakeQasmSimulator(FakeBackend):
+    """A fake simulator backend."""
+
+    def configuration(self):
+        return BackendConfiguration(
+            backend_name='fake_qasm_simulator',
+            backend_version='0.0.0',
+            n_qubits=5,
+            basis_gates=['u1', 'u2', 'u3', 'cx', 'cz', 'id', 'x', 'y', 'z',
+                         'h', 's', 'sdg', 't', 'tdg', 'ccx', 'swap',
+                         'snapshot', 'unitary'],
+            simulator=True,
+            local=True,
+            conditional=True,
+            open_pulse=False,
+            memory=True,
+            max_shots=65536,
+            gates=[GateConfig(name='TODO', parameters=[], qasm_def='TODO')],
+            coupling_map=cmap,
+        )
+
+
+class FakeTenerife(FakeBackend):
+    """A fake 5 qubit backend."""
+
+    def configuration(self):
+        cmap = [[1, 0], [2, 0], [2, 1], [3, 2], [3, 4], [4, 2]]
+
+        return BackendConfiguration(
+            backend_name='fake_tenerife',
+            backend_version='0.0.0',
+            n_qubits=5,
+            basis_gates=['u1', 'u2', 'u3', 'cx', 'id'],
+            simulator=False,
+            local=True,
+            conditional=False,
+            open_pulse=False,
+            memory=False,
+            max_shots=65536,
+            gates=[GateConfig(name='TODO', parameters=[], qasm_def='TODO')],
+            coupling_map=cmap,
+        )
+
+
+class FakeRueschlikon(FakeBackend):
+    """A fake 16 qubit backend."""
+
+    def configuration(self):
+        camp = [[1, 0], [1, 2], [2, 3], [3, 4], [3, 14], [5, 4], [6, 5],
+                [6, 7], [6, 11], [7, 10], [8, 7], [9, 8], [9, 10],
+                [11, 10], [12, 5], [12, 11], [12, 13], [13, 4],
+                [13, 14], [15, 0], [15, 2], [15, 14]]
+
+        return BackendConfiguration(
+            backend_name='fake_rueschlikon',
+            backend_version='0.0.0',
+            n_qubits=16,
+            basis_gates=['u1', 'u2', 'u3', 'cx', 'id'],
+            simulator=False,
+            local=True,
+            conditional=False,
+            open_pulse=False,
+            memory=False,
+            max_shots=65536,
+            gates=[GateConfig(name='TODO', parameters=[], qasm_def='TODO')],
+            coupling_map=cmap,
+        )
+
+
+class FakeTokyo(FakeBackend):
+    """A fake 20 qubit backend."""
+
+    def configuration(self):
+        cmap = [[0, 1], [0, 5], [1, 0], [1, 2], [1, 6], [2, 1],
+                [2, 3], [2, 6], [3, 2], [3, 8], [3, 9], [4, 8], [4, 9],
+                [5, 0], [5, 6], [5, 10], [5, 11], [6, 1], [6, 2], [6, 5],
+                [6, 7], [6, 10], [6, 11], [7, 1], [7, 6], [7, 8], [7, 12],
+                [7, 13], [8, 3], [8, 4], [8, 7], [8, 9], [8, 12], [8, 13],
+                [9, 3], [9, 4], [9, 8], [10, 5], [10, 6], [10, 11], [10, 15],
+                [11, 5], [11, 6], [11, 10], [11, 12], [11, 16], [11, 17],
+                [12, 7], [12, 8], [12, 11], [12, 13], [12, 16], [13, 7],
+                [13, 8], [13, 12], [13, 14], [13, 18], [13, 19], [14, 13],
+                [14, 18], [14, 19], [15, 10], [15, 16], [16, 11], [16, 12],
+                [16, 15], [16, 17], [17, 11], [17, 16], [18, 13], [18, 14],
+                [19, 13], [19, 14]]
+
+        return BackendConfiguration(
+            backend_name='fake_tokyo',
+            backend_version='0.0.0',
+            n_qubits=16,
+            basis_gates=['u1', 'u2', 'u3', 'cx', 'id'],
+            simulator=False,
+            local=True,
+            conditional=False,
+            open_pulse=False,
+            memory=False,
+            max_shots=65536,
+            gates=[GateConfig(name='TODO', parameters=[], qasm_def='TODO')],
+            coupling_map=cmap,
+        )
+
+
+class FakeJob(BaseJob):
+    """Fake simulator job"""
     _executor = futures.ProcessPoolExecutor()
 
     def __init__(self, fn, qobj, job_id, backend):
@@ -181,54 +297,280 @@ def new_fake_qobj():
     )
 
 
-class FakeTenerife:
-    """A fake 5 qubit backend."""
-
-    def name(self):
-        return 'fake_tenerife'
-
-    def configuration(self):
-        cmap = [[1, 0], [2, 0], [2, 1], [3, 2], [3, 4], [4, 2]]
-
-        return BackendConfiguration(
-            backend_name='fake_tenerife',
-            backend_version='0.0.0',
-            n_qubits=5,
-            basis_gates=['u1', 'u2', 'u3', 'cx', 'id'],
-            simulator=False,
-            local=True,
-            conditional=False,
-            open_pulse=False,
-            memory=False,
-            max_shots=65536,
-            gates=[GateConfig(name='TODO', parameters=[], qasm_def='TODO')],
-            coupling_map=cmap,
-        )
+def _auto_progress_api(api, interval=0.2):
+    """Progress a `BaseFakeAPI` instacn every `interval` seconds until reaching
+    the final state.
+    """
+    with suppress(BaseFakeAPI.NoMoreStatesError):
+        while True:
+            time.sleep(interval)
+            api.progress()
 
 
-class FakeRueschlikon:
-    """A fake 16 qubit backend."""
+class BaseFakeAPI():
+    """Base class for faking the IBM-Q API."""
 
-    def name(self):
-        return 'fake_rueschlikon'
+    class NoMoreStatesError(Exception):
+        """Raised when it is not possible to progress more."""
 
-    def configuration(self):
-        camp = [[1, 0], [1, 2], [2, 3], [3, 4], [3, 14], [5, 4], [6, 5],
-                [6, 7], [6, 11], [7, 10], [8, 7], [9, 8], [9, 10],
-                [11, 10], [12, 5], [12, 11], [12, 13], [13, 4],
-                [13, 14], [15, 0], [15, 2], [15, 14]]
+    _job_status = []
 
-        return BackendConfiguration(
-            backend_name='fake_rueschlikon',
-            backend_version='0.0.0',
-            n_qubits=16,
-            basis_gates=['u1', 'u2', 'u3', 'cx', 'id'],
-            simulator=False,
-            local=True,
-            conditional=False,
-            open_pulse=False,
-            memory=False,
-            max_shots=65536,
-            gates=[GateConfig(name='TODO', parameters=[], qasm_def='TODO')],
-            coupling_map=cmap,
-        )
+    _can_cancel = False
+
+    def __init__(self):
+        self._state = 0
+        self.config = {'hub': None, 'group': None, 'project': None}
+        if self._can_cancel:
+            self.config.update({
+                'hub': 'test-hub',
+                'group': 'test-group',
+                'project': 'test-project'
+            })
+
+    def get_job(self, job_id):
+        if not job_id:
+            return {'status': 'Error', 'error': 'Job ID not specified'}
+        return self._job_status[self._state]
+
+    def get_status_job(self, job_id):
+        summary_fields = ['status', 'error', 'infoQueue']
+        complete_response = self.get_job(job_id)
+        return {key: value for key, value in complete_response.items()
+                if key in summary_fields}
+
+    def run_job(self, *_args, **_kwargs):
+        time.sleep(0.2)
+        return {'id': 'TEST_ID'}
+
+    def cancel_job(self, job_id, *_args, **_kwargs):
+        if not job_id:
+            return {'status': 'Error', 'error': 'Job ID not specified'}
+        return {} if self._can_cancel else {
+            'error': 'testing fake API can not cancel'}
+
+    def progress(self):
+        if self._state == len(self._job_status) - 1:
+            raise self.NoMoreStatesError()
+        self._state += 1
+
+
+class UnknownStatusAPI(BaseFakeAPI):
+    """Class for emulating an API with unknown status codes."""
+
+    _job_status = [
+        {'status': 'UNKNOWN'}
+    ]
+
+
+class ValidatingAPI(BaseFakeAPI):
+    """Class for emulating an API with job validation."""
+
+    _job_status = [
+        {'status': 'VALIDATING'},
+        {'status': 'RUNNING'}
+    ]
+
+
+class ErrorWhileValidatingAPI(BaseFakeAPI):
+    """Class for emulating an API processing an invalid job."""
+
+    _job_status = [
+        {'status': 'VALIDATING'},
+        {'status': 'ERROR_VALIDATING_JOB'}
+    ]
+
+
+class NonQueuedAPI(BaseFakeAPI):
+    """Class for emulating a successfully-completed non-queued API."""
+
+    _job_status = [
+        {'status': 'RUNNING'},
+        {'status': 'COMPLETED', 'qasms': []}
+    ]
+
+
+class ErrorWhileCreatingAPI(BaseFakeAPI):
+    """Class emulating an API processing a job that errors while creating
+    the job.
+    """
+
+    _job_status = [
+        {'status': 'ERROR_CREATING_JOB'}
+    ]
+
+
+class ErrorWhileRunningAPI(BaseFakeAPI):
+    """Class emulating an API processing a job that errors while running."""
+
+    _job_status = [
+        {'status': 'RUNNING'},
+        {'status': 'ERROR_RUNNING_JOB', 'error': 'Error running job'}
+    ]
+
+
+class QueuedAPI(BaseFakeAPI):
+    """Class for emulating a successfully-completed queued API."""
+
+    _job_status = [
+        {'status': 'RUNNING', 'infoQueue': {'status': 'PENDING_IN_QUEUE'}},
+        {'status': 'RUNNING'},
+        {'status': 'COMPLETED'}
+    ]
+
+
+class RejectingJobAPI(BaseFakeAPI):
+    """Class for emulating an API unable of initializing."""
+
+    def run_job(self, *_args, **_kwargs):
+        return {'error': 'invalid qobj'}
+
+
+class UnavailableRunAPI(BaseFakeAPI):
+    """Class for emulating an API throwing before even initializing."""
+
+    def run_job(self, *_args, **_kwargs):
+        time.sleep(0.2)
+        raise ApiError()
+
+
+class ThrowingAPI(BaseFakeAPI):
+    """Class for emulating an API throwing in the middle of execution."""
+
+    _job_status = [
+        {'status': 'RUNNING'}
+    ]
+
+    def get_job(self, job_id):
+        raise ApiError()
+
+
+class ThrowingNonJobRelatedErrorAPI(BaseFakeAPI):
+    """Class for emulating an scenario where the job is done but the API
+    fails some times for non job-related errors.
+    """
+
+    _job_status = [
+        {'status': 'COMPLETED'}
+    ]
+
+    def __init__(self, errors_before_success=2):
+        super().__init__()
+        self._number_of_exceptions_to_throw = errors_before_success
+
+    def get_job(self, job_id):
+        if self._number_of_exceptions_to_throw != 0:
+            self._number_of_exceptions_to_throw -= 1
+            raise ApiError()
+
+        return super().get_job(job_id)
+
+
+class ThrowingGetJobAPI(BaseFakeAPI):
+    """Class for emulating an API throwing in the middle of execution. But not in
+       get_status_job() , just in get_job().
+       """
+
+    _job_status = [
+        {'status': 'COMPLETED'}
+    ]
+
+    def get_status_job(self, job_id):
+        return self._job_status[self._state]
+
+    def get_job(self, job_id):
+        raise ApiError('Unexpected error')
+
+
+class CancellableAPI(BaseFakeAPI):
+    """Class for emulating an API with cancellation."""
+
+    _job_status = [
+        {'status': 'RUNNING'},
+        {'status': 'CANCELLED'}
+    ]
+
+    _can_cancel = True
+
+
+class NonCancellableAPI(BaseFakeAPI):
+    """Class for emulating an API without cancellation running a long job."""
+
+    _job_status = [
+        {'status': 'RUNNING'},
+        {'status': 'RUNNING'},
+        {'status': 'RUNNING'}
+    ]
+
+
+class ErroredCancellationAPI(BaseFakeAPI):
+    """Class for emulating an API with cancellation but throwing while
+    trying.
+    """
+
+    _job_status = [
+        {'status': 'RUNNING'},
+        {'status': 'RUNNING'},
+        {'status': 'RUNNING'}
+    ]
+
+    _can_cancel = True
+
+    def cancel_job(self, job_id, *_args, **_kwargs):
+        return {'status': 'Error', 'error': 'test-error-while-cancelling'}
+
+
+# TODO: Remove once qobj results come by default from all the simulator
+# backends.
+class QObjResultAPI(BaseFakeAPI):
+    """Class for emulating a successfully-completed non-queued API."""
+
+    _job_status = [
+        {'status': 'RUNNING'},
+        {
+            'status': 'COMPLETED',
+            'qObjectResult': {
+                'backend_name': 'ibmqx2',
+                'backend_version': '1.1.1',
+                'job_id': 'XC1323XG2',
+                'qobj_id': 'Experiment1',
+                'success': True,
+                'status': 'COMPLETED',
+                'results': [
+                    {
+                        'header': {
+                            'name': 'Bell state',
+                            'memory_slots': 2,
+                            'creg_sizes': [['c', 2]],
+                            'clbit_labels': [['c', 0], ['c', 1]],
+                            'qubit_labels': [['q', 0], ['q', 1]]
+                        },
+                        'shots': 1024,
+                        'status': 'DONE',
+                        'success': True,
+                        'data': {
+                            'counts': {
+                                '0x0': 480, '0x3': 490, '0x1': 20, '0x2': 34
+                            }
+                        }
+                    },
+                    {
+                        'header': {
+                            'name': 'Bell state XY',
+                            'memory_slots': 2,
+                            'creg_sizes': [['c', 2]],
+                            'clbit_labels': [['c', 0], ['c', 1]],
+                            'qubit_labels': [['q', 0], ['q', 1]]
+                        },
+                        'shots': 1024,
+                        'status': 'DONE',
+                        'success': True,
+                        'data': {
+                            'counts': {
+                                '0x0': 29, '0x3': 15, '0x1': 510, '0x2': 480
+                            }
+                        }
+                    }
+                ]
+            }
+        }
+    ]
